refactor(callbacks): extract Alerts tab helpers to _callbacks_alerts.py
Step 4 of the callbacks.py decomposition (issue #87). Continues the
per-tab split established by #98 (shared), #99 (US Grid), #100 (Models).

## What moves

``_alerts_tab_from_redis`` — the 200-line Redis fast path that builds
the entire Alerts / Extreme Events tab — moves out of callbacks.py into
``components/_callbacks_alerts.py``.

The Alerts tab is structurally Redis-only: every view it renders comes
from the ``wattcast:alerts:{region}`` payload that the scoring job
writes each hour. There's no inline-compute fallback because alert
detection itself runs in the scoring job, not on the web. That makes
this module small (~200 lines) and self-contained — only one function,
no register_callbacks v1 fallback to keep in sync.

## Public-import surface preserved

``components/callbacks.py`` re-imports ``_alerts_tab_from_redis`` by
name. ``from components.callbacks import _alerts_tab_from_redis`` in
tests + the ``register_callbacks`` wiring continues to resolve. No
caller-side changes.

## Test patches updated

7 ``@patch("components.callbacks.redis_get")`` decorators inside
``TestAlertsTabFromRedis`` rewired to ``components._callbacks_alerts``
— standard "patch where it's used, not where it's defined."

## Size impact

callbacks.py: 6,183 → 5,978 (–205 lines, crossing below 6K for first time)
_callbacks_alerts.py: new, 260 lines (function + module docstring + __all__)

Cumulative progress on the decomposition:
* PR #98 (shared):   7,167 → 7,051 (–116)
* PR #99 (us_grid):  7,051 → 6,389 (–662)
* PR #100 (models):  6,389 → 6,183 (–206)
* this PR (alerts):  6,183 → 5,978 (–205)

## Tests

All 1,568 unit tests pass locally. Ruff clean.
